547 remote

+++ lib/Ravada.pm
@@ -11,6 +11,7 @@ use DBIx::Connector;
 use File::Copy;
 use Hash::Util qw(lock_hash);
 use Moose;
+use Parallel::ForkManager;
 use POSIX qw(WNOHANG);
 use YAML;
 
@@ -70,27 +71,13 @@ our $CAN_FORK = 1;
 our $CAN_LXC = 0;
 
 # Seconds to wait for other long process
-our $SECONDS_WAIT_CHILDREN = 2;
-# Limit for long processes
-our $LIMIT_PROCESS = 2;
-our $LIMIT_HUGE_PROCESS = 1;
+our $SECONDS_WAIT_CHILDREN = 5;
 
 our $DIR_SQL = "sql/mysql";
 $DIR_SQL = "/usr/share/doc/ravada/sql/mysql" if ! -e $DIR_SQL;
 
-# LONG commands take long
-our %HUGE_COMMAND = map { $_ => 1 } qw(download);
-our %LONG_COMMAND =  map { $_ => 1 } (qw(prepare_base remove_base screenshot set_base_vm ), keys %HUGE_COMMAND);
-
 our $USER_DAEMON;
 our $USER_DAEMON_NAME = 'daemon';
-
-has 'vm' => (
-          is => 'ro'
-        ,isa => 'ArrayRef'
-       ,lazy => 1
-     , builder => '_create_vm'
-);
 
 has 'connector' => (
         is => 'rw'
@@ -852,8 +839,6 @@ sub _init_config {
 
     $CONFIG->{vm} = [] if !$CONFIG->{vm};
 
-    $LIMIT_PROCESS = $CONFIG->{limit_process} 
-        if $CONFIG->{limit_process} && $CONFIG->{limit_process}>1;
 #    $CONNECTOR = ( $connector or _connect_dbh());
 
     _init_config_vm();
@@ -1059,6 +1044,7 @@ sub create_domain {
 
     my $domain;
     eval { $domain = $vm->create_domain(@create_args) };
+    die $@ if $@ && !$request;
     my $error = $@;
     $request->error($error) if $request && $error;
     if ($error =~ /has \d+ requests/) {
@@ -1427,13 +1413,13 @@ sub remove_volume {
 
 }
 
-=head2 clean_killed_requests
+=head2 clean_old_requests
 
-Before processing requests, old killed requests must be cleaned.
+Before processing requests, old requests must be cleaned.
 
 =cut
 
-sub clean_killed_requests {
+sub clean_old_requests {
     my $self = shift;
     my $sth = $CONNECTOR->dbh->prepare("SELECT id FROM requests "
         ." WHERE status <> 'done' AND STATUS <> 'requested'"
@@ -1444,6 +1430,8 @@ sub clean_killed_requests {
         $req->status("done","Killed ".$req->command." before completion");
     }
 
+    $self->_clean_requests('refresh_vms');
+    $self->_clean_requests('cleanup');
 }
 
 =head2 process_requests
@@ -1458,10 +1446,11 @@ sub process_requests {
     my $self = shift;
     my $debug = shift;
     my $dont_fork = shift;
-    my $long_commands = (shift or 0);
-    my $short_commands = (shift or 0);
+    my $request_type = ( shift or 'all');
+    confess "ERROR: Request type '$request_type' unknown, it must be long, huge, all"
+            ." or priority"
+        if $request_type !~ /^(long|huge|priority|all)$/;
 
-    $self->_wait_pids_nohang();
     $self->_kill_stale_process();
 
     my $sth = $CONNECTOR->dbh->prepare("SELECT id,id_domain FROM requests "
@@ -1472,36 +1461,27 @@ sub process_requests {
     );
     $sth->execute(time);
 
-    my $debug_type = '';
-    $debug_type = 'long' if $long_commands;
-    $debug_type = 'short' if $short_commands || !$long_commands;
-    $debug_type = 'all' if $long_commands && $short_commands;
-
     while (my ($id_request,$id_domain)= $sth->fetchrow) {
-        my $req = Ravada::Request->open($id_request);
+        my $req;
+        eval { $req = Ravada::Request->open($id_request) };
+        next if $@ && $@ =~ /I can't find/;
+        warn $@ if $@;
+        next if !$req;
 
-        if ( ($long_commands && 
-                (!$short_commands && !$LONG_COMMAND{$req->command}))
-            ||(!$long_commands && $LONG_COMMAND{$req->command})
-        ) {
-            warn "[$debug_type,$long_commands,$short_commands] $$ skipping request "
-                .$req->command
-                    if $debug || $DEBUG;
-            next;
-        }
+        next if $request_type ne 'all' && $req->type ne $request_type;
+
         next if $req->command !~ /shutdown/i
             && $self->_domain_working($id_domain, $id_request);
 
-        warn "[$debug_type] $$ executing request ".$req->id." ".$req->status()." "
+        warn "[$request_type] $$ executing request ".$req->id." ".$req->status()." "
             .$req->command
             ." ".Dumper($req->args) if $DEBUG || $debug;
 
         my ($n_retry) = $req->status() =~ /retry (\d+)/;
         $n_retry = 0 if !$n_retry;
-        my $err = $self->_execute($req, $dont_fork);
-        $req->error($err)   if $err;
+
+        $self->_execute($req, $dont_fork);
 #        $req->status("done") if $req->status() !~ /retry/;
-        $self->_cmd_refresh_vms();
         next if !$DEBUG && !$debug;
 
         warn "req ".$req->id." , command: ".$req->command." , status: ".$req->status()
@@ -1509,7 +1489,6 @@ sub process_requests {
 
     }
     $sth->finish;
-
 }
 
 =head2 process_long_requests
@@ -1522,7 +1501,7 @@ sub process_long_requests {
     my $self = shift;
     my ($debug,$dont_fork) = @_;
 
-    return $self->process_requests($debug, $dont_fork, 1);
+    return $self->process_requests($debug, $dont_fork, 'long');
 }
 
 =head2 process_all_requests
@@ -1536,10 +1515,15 @@ sub process_all_requests {
     my $self = shift;
     my ($debug,$dont_fork) = @_;
 
-    $self->process_requests($debug, $dont_fork,1,1);
+    $self->process_requests($debug, $dont_fork,'all');
 
 }
 
+sub process_priority_requests($self, $debug=0, $dont_fork=0) {
+
+    $self->process_requests($debug, $dont_fork,'priority');
+
+}
 
 sub _kill_stale_process($self) {
     my $sth = $CONNECTOR->dbh->prepare(
@@ -1553,7 +1537,12 @@ sub _kill_stale_process($self) {
     );
     $sth->execute(time - 60 );
     while (my ($pid, $command, $start_time) = $sth->fetchrow) {
-        warn "Killing $command stale for ".time - $start_time." seconds\n";
+        if ($pid == $$ ) {
+            warn "HOLY COW! I should kill pid $pid stale for ".(time - $start_time)
+                ." seconds, but I won't because it is myself";
+            next;
+        }
+        warn "Killing $command stale for ".(time - $start_time)." seconds\n";
         kill (15,$pid);
     }
     $sth->finish;
@@ -1593,13 +1582,14 @@ sub _process_all_requests_dont_fork {
     my $self = shift;
     my $debug = shift;
 
-    return $self->process_requests($debug,1, 1, 1);
+    return $self->process_requests($debug,1, 'all');
 }
 
 sub _process_requests_dont_fork {
     my $self = shift;
     my $debug = shift;
-    return $self->process_requests($debug, 1);
+    return $self->process_requests($debug, 'priority');
+    return $self->process_requests($debug, 'long');
 }
 
 =head2 list_vm_types
@@ -1632,31 +1622,37 @@ sub _execute {
     $request->pid($$);
     $request->start_time(time);
     $request->error('');
-    if ($dont_fork || !$CAN_FORK || !$LONG_COMMAND{$request->command}) {
+    if ($dont_fork || !$CAN_FORK) {
 
         eval { $sub->($self,$request) };
         my $err = ($@ or '');
         $request->status('done') if $request->status() ne 'done'
                                     && $request->status !~ /retry/;
         $request->error($err) if $err;
-        return $err;
+        return;
     }
 
-    $self->_wait_pids_nohang();
-    return if $self->_wait_children($request);
+    if ( $self->_wait_requests($request) ) {
+         $request->status("requested","Server loaded, queuing request");
+         return;
+     }
 
-    $request->status('working');
-    warn $request->command." forking";
-    my $pid = fork();
+    $request->status('working','');
+    if (!$self->{fork_manager}) {
+        my $fm = Parallel::ForkManager->new($request->requests_limit('priority'));
+        $self->{fork_manager} = $fm;
+    }
+    my $pid = $self->{fork_manager}->start;
     die "I can't fork" if !defined $pid;
+
     if ( $pid == 0 ) {
         $self->_do_execute_command($sub, $request);
-    } else {
-        $request->pid($pid);
-        $self->_add_pid($pid, $request->id);
+        $self->{fork_manager}->finish; # Terminates the child process
+        $request->status('done');
+        exit;
     }
-#    $self->_connect_vm_kvm();
-    return '';
+    $request->pid($pid);
+    $self->{fork_manager}->reap_finished_children;
 }
 
 sub _do_execute_command {
@@ -1673,12 +1669,10 @@ sub _do_execute_command {
 #    }
 
     eval {
-        $self->_connect_vm();
         $sub->($self,$request);
-        $self->_disconnect_vm();
     };
     my $err = ( $@ or '');
-    $request->error($err);
+    $request->error($err)   if $err;
     $request->status('done') 
         if $request->status() ne 'done'
             && $request->status() !~ /^retry/i;
@@ -1744,7 +1738,7 @@ sub _cmd_create{
     my $request = shift;
 
     $request->status('creating domain');
-    warn "$$ creating domain"   if $DEBUG;
+    warn "$$ creating domain ".Dumper($request->args)   if $DEBUG;
     my $domain;
 
     $domain = $self->create_domain(%{$request->args},request => $request);
@@ -1757,37 +1751,32 @@ sub _cmd_create{
             .$request->args('name')."</a>"
             ." created."
         ;
+        warn $msg if $DEBUG;
         $request->status('done',$msg);
     }
 
 
 }
 
-sub _wait_children {
+sub _wait_requests {
     my $self = shift;
     my $req = shift or confess "Missing request";
 
+    # don't wait for priority requests
+    return if $req->type eq 'priority';
+
     my $try = 0;
     for ( 1 .. $SECONDS_WAIT_CHILDREN ) {
-        my $n_pids = scalar keys %{$self->{pids}};
 
         my $msg;
-        if ($HUGE_COMMAND{$req->command}) {
-            if ( $n_pids < $LIMIT_HUGE_PROCESS) {
-                $msg = $req->id." ".$req->command
+
+        my $n_pids = $req->count_requests();
+
+        $msg = $req->command
                 ." waiting for processes to finish $n_pids"
-                ." of $LIMIT_HUGE_PROCESS ";
-                warn $msg if $DEBUG;
-                return;
-            }
-        } elsif ( $n_pids < $LIMIT_PROCESS) {
-            $msg = $req->id." ".$req->command
-                ." waiting for processes to finish $n_pids"
-                ." of $LIMIT_PROCESS ";
-            warn $msg if $DEBUG;
-            return;
-        }
-        $self->_wait_pids_nohang();
+                ." of ".$req->requests_limit;
+        return if $n_pids < $req->requests_limit();
+        return 1 if $n_pids > $req->requests_limit + 2;
         sleep 1;
 
         next if $try++;
@@ -1795,7 +1784,7 @@ sub _wait_children {
         $req->error($msg);
         $req->status('waiting') if $req->status() !~ 'waiting';
     }
-    return scalar keys %{$self->{pids}};
+    return 1;
 }
 
 sub _wait_pids_nohang {
@@ -1804,9 +1793,9 @@ sub _wait_pids_nohang {
 
     for my $pid ( keys %{$self->{pids}}) {
         my $kid = waitpid($pid , WNOHANG);
-        next if !$kid || $kid == -1;
-        $self->_set_req_done($kid);
-        $self->_delete_pid($kid);
+        next if !$kid;
+        $self->_set_req_done($pid);
+        $self->_delete_pid($pid);
     }
 
 }
@@ -1841,22 +1830,6 @@ sub _wait_pids {
         $self->_delete_pid($kid);
         return if $kid  == $pid;
     }
-}
-
-sub _add_pid {
-    my $self = shift;
-    my $pid = shift;
-    my $id_req = shift;
-
-    $self->{pids}->{$pid} = $id_req;
-
-}
-
-sub _delete_pid {
-    my $self = shift;
-    my $pid = shift;
-
-    delete $self->{pids}->{$pid};
 }
 
 sub _cmd_remove {
@@ -1945,8 +1918,8 @@ sub _cmd_start {
 
     my $name = $request->args('name');
 
-    my $domain = $self->search_domain($name);
-    die "Unknown domain '$name'" if !$domain;
+    my $domain = $self->search_domain($name)
+        or die "Unknown domain $name ";
 
     my $uid = $request->args('uid');
     my $user = Ravada::Auth::SQL->search_by_id($uid);
@@ -2006,7 +1979,9 @@ sub _cmd_hybernate {
     my $id_domain = $request->id_domain or confess "Missing request id_domain";
 
     my $user = Ravada::Auth::SQL->search_by_id( $uid);
-    my $domain = $self->search_domain_by_id($id_domain);
+
+    warn "open $id_domain";
+    my $domain = Ravada::Domain->open($id_domain);
 
     die "Unknown domain id '$id_domain'\n" if !$domain;
 
@@ -2045,16 +2020,22 @@ sub _cmd_shutdown {
     my $name = $request->defined_arg('name');
     my $id_domain = $request->defined_arg('id_domain');
     my $timeout = ($request->args('timeout') or 60);
+    my $id_vm = $request->defined_arg('id_vm');
 
     confess "ERROR: Missing id_domain or name" if !$id_domain && !$name;
 
     my $domain;
     if ($name) {
-    $domain = $self->search_domain($name);
-    die "Unknown domain '$name'\n" if !$domain;
+        if ($id_vm) {
+            my $vm = Ravada::VM->open($id_vm);
+            $domain = $vm->search_domain($name);
+        } else {
+            $domain = $self->search_domain($name);
+        }
+        die "Unknown domain '$name'\n" if !$domain;
     }
     if ($id_domain) {
-        my $domain2 = $self->search_domain_by_id($id_domain);
+        my $domain2 = Ravada::Domain->open(id => $id_domain, id_vm => $id_vm);
         die "ERROR: Domain $id_domain is ".$domain2->name." not $name."
             if $domain && $domain->name ne $domain2->name;
         $domain = $domain2;
@@ -2151,23 +2132,24 @@ sub _cmd_refresh_vms($self, $request=undef) {
     my ($active_domain, $active_vm) = $self->_refresh_active_domains($request);
     $self->_refresh_down_domains($active_domain, $active_vm);
 
-    $self->_clean_refresh_vms_requests($request);
+    $self->_clean_requests('refresh_vms', $request);
 }
 
-sub _clean_refresh_vms_requests($self, $request) {
+sub _clean_requests($self, $command, $request=undef) {
     my $query = "DELETE FROM requests "
-        ." WHERE command='refresh_vm' "
+        ." WHERE command=? "
         ."   AND status='requested'";
 
     if ($request) {
+        confess "Wrong request" if !ref($request) || ref($request) !~ /Request/;
         $query .= "   AND id <> ?";
     }
     my $sth = $CONNECTOR->dbh->prepare($query);
 
     if ($request) {
-        $sth->execute($request->id);
+        $sth->execute($command, $request->id);
     } else {
-        $sth->execute();
+        $sth->execute($command);
     }
 }
 
@@ -2253,6 +2235,11 @@ sub _cmd_set_base_vm {
      );
 }
 
+sub _cmd_cleanup($self, $request) {
+    $self->enforce_limits( request => $request);
+    $self->_clean_requests('cleanup', $request);
+}
+
 sub _req_method {
     my $self = shift;
     my  $cmd = shift;
@@ -2265,6 +2252,7 @@ sub _req_method {
         ,create => \&_cmd_create
         ,remove => \&_cmd_remove
         ,resume => \&_cmd_resume
+       ,cleanup => \&_cmd_cleanup
       ,download => \&_cmd_download
       ,shutdown => \&_cmd_shutdown
      ,hybernate => \&_cmd_hybernate
@@ -2272,6 +2260,7 @@ sub _req_method {
     ,domdisplay => \&_cmd_domdisplay
     ,screenshot => \&_cmd_screenshot
     ,copy_screenshot => \&_cmd_copy_screenshot
+   ,cmd_cleanup => \&_cmd_cleanup
    ,remove_base => \&_cmd_remove_base
    ,set_base_vm => \&_cmd_set_base_vm
   ,ping_backend => \&_cmd_ping_backend
@@ -2321,15 +2310,36 @@ sub search_vm {
         return Ravada::VM::Void->new(host => $host);
     }
 
-    my @vms;
-    eval { @vms = @{$self->vm} };
-    return if $@ && $@ =~ /No VMs found/i;
-    die $@ if $@;
+    my $sth = $CONNECTOR->dbh->prepare(
+        "SELECT id FROM vms "
+        ." WHERE vm_type = ? "
+        ."   AND hostname=?"
+    );
+    $sth->execute($type, $host);
+    my ($id) = $sth->fetchrow();
+    return Ravada::VM->open($id)    if $id;
+    return if $host ne 'localhost';
 
-    for my $vm (@vms) {
+    my $vms = $self->_create_vm();
+
+    for my $vm (@$vms) {
         return $vm if ref($vm) eq $class && $vm->host eq $host;
     }
     return;
+}
+
+sub vm($self) {
+    my $sth = $CONNECTOR->dbh->prepare(
+        "SELECT id FROM vms "
+    );
+    $sth->execute();
+    my @vms;
+    while ( my ($id) = $sth->fetchrow()) {
+        push @vms, ( Ravada::VM->open($id));
+    };
+    return [@vms] if @vms;
+    return $self->_create_vms();
+
 }
 
 =head2 import_domain
@@ -2383,9 +2393,11 @@ sub enforce_limits {
 
 sub _enforce_limits_active {
     my $self = shift;
+    confess "Even size args" if scalar(@_) % 2;
     my %args = @_;
 
     my $timeout = (delete $args{timeout} or 10);
+    my $request = delete $args{request};
 
     confess "ERROR: Unknown arguments ".join(",",sort keys %args)
         if keys %args;
@@ -2396,6 +2408,9 @@ sub _enforce_limits_active {
     }
     for my $id_user(keys %domains) {
         next if scalar @{$domains{$id_user}}<2;
+
+        my $user = Ravada::Auth::SQL->search_by_id($id_user);
+        next if $user->is_admin();
 
         my @domains_user = sort { $a->start_time <=> $b->start_time }
                         @{$domains{$id_user}};
